Fix for big-endian machines.
The problem is that the card provides a big-endian view on its
registers, which the driver uses on big-endian machines.  However, the
driver uses MMIO_OUT32() and friends to access these registers.  And
on sparc64, these macros do byte swapping themselves.  The net result
of course is that no byte swapping gets done at all, and things fail
miserably.

The fix is to make the macros not do byte swapping, which can be done
by defining SPARC_MMIO_IS_BE.  There is a similar define for powerpc
as well.  I assume these defines were accidentally dropped in the
autoconfiscation of Xorg.

Other people came up with the same conclusion see:

  http://www.mail-archive.com/opensolaris-discuss@opensolaris.org/msg20874.html

and

  https://bugs.freedesktop.org/show_bug.cgi?id=10742

Signed-off-by: Matthieu Herrb <matthieu.herrb@laas.fr>