Only map a 64k block of registers.
The driver tries to map a block of registers that is too large.  The
mmio registers consist of two blocks of 64k.  The first 64k provide a
little-endian view, the second 64k provide a big-endian view.  However
the driver always tries to map 128k.  And mapping 128k from the offset
where the big-endian view starts will fail on OpenBSD.

This changes things such that the driver maps just 64k.  There is no
reason to map the big-endian view on little-endian machines.

Signed-off-by: Matthieu Herrb <matthieu.herrb@laas.fr>