Make pm2's xv driver collect options like all other drivers.
The current method of argument collection is to collect options from different
ports of a VideoAdaptor record. Specifically, the ports had to be named
'Input' for input options, and 'Output' for output options.

This resulted in three groups of options, requiring people to know what
VideoAdaptor does, both of which were not documented in the man page. If you
forgot to create a VideoAdaptor port, then the xv driver would just not work.

This patch makes the xv driver collect options from the screen, like every
single other driver. Input and Output prefixes are used for options where the
input and output ports have the same args (FramesPerSec, for example).
Documentation added for the change.

This is a step toward getting rid of VideoAdaptor, since only glint uses it
(and is probably the only one to have used it).

v2: Complain about Xv driver failing to load only if the user wanted Xv.
Don't use pGlint->Options, since glint is still initializing.

Signed-off-by: Jesse Adkins <jesserayadkins@gmail.com>
Signed-off-by: Matt Turner <mattst88@gmail.com>