Fix uploading YV12 data to texture buffer on BE machines
On BE machines various hardware byteswapping options are used for the
framebuffer aperture.  Which option gets used depends on the depth of the
framebuffer.  Uploading YV12 data to the texture buffer is done through the
same aperture, but is always done in 32-bit wide units.  Therefore the code
that does the uploading needs to take into account the byteswapping done by
the hardware.  For 32bpp modes we can use the same code as on LE machines,
but 16bpp and 8bpp modes need their own versions.

Signed-off-by: Mark Kettenis <kettenis@openbsd.org>
Signed-off-by: Matt Turner <mattst88@gmail.com>